--- bin/node-template/.editorconfig
@@ -1,16 +1,0 @@
-root = true
-
-[*]
-indent_style=space
-indent_size=2
-tab_width=2
-end_of_line=lf
-charset=utf-8
-trim_trailing_whitespace=true
-insert_final_newline = true
-
-[*.{rs,toml}]
-indent_style=tab
-indent_size=tab
-tab_width=4
-max_line_length=100
