--- bin/node-template/node/build.rs
@@ -1,7 +1,0 @@
-use substrate_build_script_utils::{generate_cargo_keys, rerun_if_git_head_changed};
-
-fn main() {
-	generate_cargo_keys();
-
-	rerun_if_git_head_changed();
-}
