--- bin/node-template/node/src/main.rs
@@ -1,13 +1,0 @@
-//! Substrate Node Template CLI library.
-#![warn(missing_docs)]
-
-mod chain_spec;
-#[macro_use]
-mod service;
-mod cli;
-mod command;
-mod rpc;
-
-fn main() -> sc_cli::Result<()> {
-	command::run()
-}
