--- bin/node-template/pallets/template/README.md
@@ -1,1 +1,0 @@
-License: Unlicense
