--- bin/node-template/pallets/template/rpc/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "pallet-cherry-rpc"
-version = "4.0.0-dev"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-license = "Apache-2.0"
-homepage = "https://substrate.dev"
-repository = "https://github.com/paritytech/substrate/"
-description = "RPC interface for the  cherrystorage pallet."
-readme = "README.md"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "2.0.0" }
-jsonrpc-core = "18.0.0"
-jsonrpc-core-client = "18.0.0"
-jsonrpc-derive = "18.0.0"
-
-sp-api = { version = "4.0.0-dev", path = "../../../../../primitives/api" }
-sp-blockchain = { version = "4.0.0-dev", path = "../../../../../primitives/blockchain" }
-sp-core = { version = "4.0.0-dev", path = "../../../../../primitives/core" }
-sp-rpc = { version = "4.0.0-dev", path = "../../../../../primitives/rpc" }
-sp-runtime = { version = "4.0.0-dev", path = "../../../../../primitives/runtime" }
-pallet-cherry-rpc-runtime-api = { version = "4.0.0-dev", path = "./runtime-api" }
-
