--- bin/node-template/pallets/template/rpc/README.md
@@ -1,3 +1,0 @@
-RPC interface for the pallet-template pallet.
-
-License: Apache-2.0
