--- bin/node-template/pallets/template/rpc/runtime-api/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "pallet-cherry-rpc-runtime-api"
-version = "4.0.0-dev"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-license = "Apache-2.0"
-homepage = "https://substrate.dev"
-repository = "https://github.com/paritytech/substrate/"
-description = "RPC runtime API for transaction payment FRAME pallet"
-readme = "README.md"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
-sp-api = { version = "4.0.0-dev", default-features = false, path = "../../../../../../primitives/api" }
-sp-core = { version = "4.0.0-dev", default-features = false, path = "../../../../../../primitives/core" }
-sp-runtime = { version = "4.0.0-dev", default-features = false, path = "../../../../../../primitives/runtime" }
-pallet-cherry= { version = "3.0.0", default-features = false, path = "../../../template" }
-
-[features]
-default = ["std"]
-std = [
-	"codec/std",
-	"sp-api/std",
-	"sp-core/std",
-	"sp-runtime/std",
-	"pallet-cherry/std",
-]
