--- bin/node-template/pallets/template/rpc/runtime-api/README.md
@@ -1,3 +1,0 @@
-Runtime API definition for pallet-template pallet
-
-License: Apache-2.0
