--- bin/node-template/scripts/docker_run.sh
@@ -1,10 +1,0 @@
-#!/usr/bin/env bash
-# This script is meant to be run on Unix/Linux based systems
-set -e
-
-echo "*** Start Substrate node template ***"
-
-cd $(dirname ${BASH_SOURCE[0]})/..
-
-docker-compose down --remove-orphans
-docker-compose run --rm --service-ports dev $@
