--- bin/node-template/scripts/init.sh
@@ -1,12 +1,0 @@
-#!/usr/bin/env bash
-# This script is meant to be run on Unix/Linux based systems
-set -e
-
-echo "*** Initializing WASM build environment"
-
-if [ -z $CI_PROJECT_NAME ] ; then
-   rustup update nightly
-   rustup update stable
-fi
-
-rustup target add wasm32-unknown-unknown --toolchain nightly
