+++ bin/node/runtime/Cargo.toml
@@ -98,6 +98,8 @@ pallet-transaction-storage = { version = "4.0.0-dev", default-features = false, 
 pallet-uniques = { version = "4.0.0-dev", default-features = false, path = "../../../frame/uniques" }
 pallet-vesting = { version = "4.0.0-dev", default-features = false, path = "../../../frame/vesting" }
 
+pallet-ipfs = { version = "4.0.0-dev", default-features = false, path = "../pallets/pallet-ipfs" }
+
 [build-dependencies]
 substrate-wasm-builder = { version = "5.0.0-dev", path = "../../../utils/wasm-builder" }
